tests: Fix the `docker run` command (for test validator) on macOS
`--net=host` option works as expected only on Linux, where Docker
is used without any virtualization laver. On macOS, Docker runs
inside a VM, so we need to map the ports explicitly. That way,
Docker on macOS knows which ports to map first from the container
to the VM, then from the VM to the host system.